+++ src/main.ts
@@ -3,8 +3,8 @@ import express from "express"
 import dotenv from "dotenv"
 
 import { RequestFilter } from "./types"
-import { parseStatement } from "./parseStatement"
-import { handleStatement } from "./handleStatement"
+import { parseStatement } from "./statement_processing/parse"
+import { handleStatement } from "./statement_processing/handle"
 
 dotenv.config()
 
@@ -14,10 +14,10 @@ const statementPath = path.resolve(process.env.STATEMENT_PATH || "")
 const app = express()
 
 app.get("/**", async (req, res) => {
-    const { from, to, symbol } = req.query as RequestFilter
+    const { from, to, symbol, currency } = req.query as RequestFilter
 
     const statement = await parseStatement(statementPath)
-    const report = handleStatement(statement, { from, to, symbol })
+    const report = handleStatement(statement, { from, to, symbol, currency })
     res.json(report)
 })